+++ m8tes/_client.py
@@ -28,6 +28,7 @@ from ._resources import (
     Settings,
     Skills,
     Tasks,
+    Triggers,
     Users,
     Webhooks,
 )
@@ -77,6 +78,7 @@ class M8tes:
         self.github_app = GitHubApp(self._http)
         self.runs = Runs(self._http)
         self.tasks = Tasks(self._http)
+        self.triggers = Triggers(self._http)
         self.apps = Apps(self._http)
         self.built_in_tools = BuiltInTools(self._http)
         self.mcp_servers = McpServers(self._http)